the jits
Made it so that platforms are now semi-solids and are interactable, fixed the player movement to be more fluid and detailed as the one previous was clunky, went through possible 8-bit music ideas and made sure to add them to a playlist for later and began experimenting with different songs. Trees are foreground objects.