ty dolla $ign
I made it so that when you defeat big foot in level one and toss him off of the level, you are taken to level two automatically. I added the you lose screen as well for level two with all working buttons and a restart option for the level; instead of having to restart the entire game from the start. Basically a lot of testing today.